Remove MasterService class
This has been deprecated and is replaced by ClusterManagerService.
