+++ server/src/main/java/org/opensearch/cluster/service/ClusterService.java
@@ -41,7 +41,6 @@ import org.opensearch.cluster.ClusterStateTaskConfig;
 import org.opensearch.cluster.ClusterStateTaskExecutor;
 import org.opensearch.cluster.ClusterStateTaskListener;
 import org.opensearch.cluster.LocalNodeClusterManagerListener;
-import org.opensearch.cluster.LocalNodeMasterListener;
 import org.opensearch.cluster.NodeConnectionsService;
 import org.opensearch.cluster.node.DiscoveryNode;
 import org.opensearch.cluster.routing.OperationRouting;
@@ -247,22 +246,7 @@ public class ClusterService extends AbstractLifecycleComponent {
         clusterApplierService.addLocalNodeClusterManagerListener(listener);
     }
 
-    /**
-     * Add a listener for on/off local node cluster-manager events
-     * @deprecated As of 2.2, because supporting inclusive language, replaced by {@link #addLocalNodeClusterManagerListener}
-     */
-    @Deprecated
-    public void addLocalNodeMasterListener(LocalNodeMasterListener listener) {
-        addLocalNodeClusterManagerListener(listener);
-    }
-
     public ClusterManagerService getClusterManagerService() {
-        return clusterManagerService;
-    }
-
-    /** @deprecated As of 2.2, because supporting inclusive language, replaced by {@link #getClusterManagerService()} */
-    @Deprecated
-    public MasterService getMasterService() {
         return clusterManagerService;
     }
 
@@ -289,12 +273,6 @@ public class ClusterService extends AbstractLifecycleComponent {
             || Thread.currentThread().getName().contains(ClusterManagerService.CLUSTER_MANAGER_UPDATE_THREAD_NAME)
             : "not called from the master/cluster state update thread";
         return true;
-    }
-
-    /** @deprecated As of 2.2, because supporting inclusive language, replaced by {@link #assertClusterOrClusterManagerStateThread} */
-    @Deprecated
-    public static boolean assertClusterOrMasterStateThread() {
-        return assertClusterOrClusterManagerStateThread();
     }
 
     public ClusterName getClusterName() {
